Oops. Part1 was only working after refactor because the default semantics matched native int. Changed RouteValue cast operator to explicit, and implemented relational operators properly

+++ AdventOfCode/Part9/Salesman.h
@@ -58,7 +58,7 @@ namespace Part9
                 remainingCities.push_back(ii);
             }
 
-            return currentBest;
+            return (int)currentBest;
         }
     };
 }
